+++ src/qibo/__init__.py
@@ -1,5 +1,6 @@
 __version__ = "0.1.3-dev"
-from qibo.config import set_precision, set_backend, set_device, set_threads, get_threads, get_backend, get_precision, get_device, matrices, K
+from qibo.config import set_precision, set_backend, set_device, set_threads, get_threads, get_backend, get_precision, get_device
+from qibo.config import numpy, matrices, K
 from qibo import callbacks
 from qibo import evolution
 from qibo import models
